# server/openapi.yaml
openapi: 3.1.0
info:
  title: WealthGenie Public API
  version: 2.0.0
  description: Canonical contract for the supported public Express API. Internal metrics and MCP routes are intentionally excluded.
servers:
  - { url: 'http://localhost:5000', description: Local development }
  - { url: '/', description: Same-origin production gateway }
paths:
...
  /api/profile/build:
    post:
      operationId: buildProfile
      tags: [Profiles]
      security: [{ CookieAuth: [] }, { BearerAuth: [] }]
      x-csrf-protection: cookie-session
      x-idempotency-key: optional
      requestBody: { $ref: '#/components/requestBodies/ProfileCreate' }
      responses: { '201': { $ref: '#/components/responses/Profile' }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '409': { $ref: '#/components/responses/Conflict' } }
  /api/profile/precompute:
    post:
      operationId: precomputeProfileRecommendation
      tags: [Profiles]
      security: [{ CookieAuth: [] }, { BearerAuth: [] }]
      x-csrf-protection: cookie-session
      x-idempotency-key: optional
      description: Server-side, short-lived non-authoritative recommendation precomputation. It does not persist profile, recommendation, or audit records.
      requestBody: { $ref: '#/components/requestBodies/ProfileCreate' }
      responses: { '200': { $ref: '#/components/responses/ProfileCandidate' }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '503': { $ref: '#/components/responses/Unavailable' } }
  /api/profile/complete:
    post:
      operationId: completeFinancialProfile
      tags: [Profiles]
      security: [{ CookieAuth: [] }, { BearerAuth: [] }]
      x-csrf-protection: cookie-session
      x-idempotency-key: required
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      description: Authoritative profile completion boundary. Commits the profile, recommendation, audit record, and idempotency result atomically. Advisory prose remains deferred.
      requestBody: { $ref: '#/components/requestBodies/ProfileCompletion' }
      responses: { '200': { $ref: '#/components/responses/ProfileCompletion' }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '409': { $ref: '#/components/responses/Conflict' }, '503': { $ref: '#/components/responses/Unavailable' } }
  /api/profile/current:
    get:
      operationId: getCurrentProfile
      tags: [Profiles]
      security: [{ CookieAuth: [] }, { BearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Profile' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' } }
...
  /api/profile/{profileId}:
    put:
      operationId: updateProfile
      tags: [Profiles]
      security: [{ CookieAuth: [] }, { BearerAuth: [] }]
      x-csrf-protection: cookie-session
      parameters: [{ $ref: '#/components/parameters/ProfileId' }]
      requestBody: { $ref: '#/components/requestBodies/ProfileUpdate' }
      responses: { '200': { $ref: '#/components/responses/Profile' }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
...
components:
  securitySchemes:
    CookieAuth: { type: apiKey, in: cookie, name: wg_session, description: Secure HttpOnly browser session cookie. }
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Explicitly configured non-browser compatibility authentication. }
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 128 } }
    ProfileId: { name: profileId, in: path, required: true, schema: { type: string, pattern: '^[0-9a-fA-F]{24}$' } }
...
  requestBodies:
...
    ProfileCreate: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FinancialProfileInput' } } } }
    ProfileCompletion: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FinancialProfileCompletion' } } } }
    ProfileUpdate: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FinancialProfileUpdate' } } } }
...
  responses:
...
    Profile: { description: Authoritative financial profile, content: { application/json: { schema: { $ref: '#/components/schemas/FinancialProfile' } } } }
    ProfileCandidate: { description: Short-lived non-authoritative precomputation handle, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileCandidate' } } } }
    ProfileCompletion: { description: Atomically committed profile and recommendation, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileCompletion' } } } }
...
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Authentication required or invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unavailable: { description: Service or dependency unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error:
      type: object
      required: [error, message, code, request_id]
      properties:
        error: { type: string, description: Backward-compatible summary }
        message: { type: string }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        request_id: { type: [string, 'null'] }
        details: { oneOf: [{ type: array, items: { type: string } }, { type: object, additionalProperties: true }] }
...
    FinancialProfileInput:
      type: object
      unevaluatedProperties: false
      required: [monthly_take_home, monthly_savings, age, risk_tolerance, investment_goals, investment_horizon_years]
      properties:
        monthly_take_home: { type: number, exclusiveMinimum: 0, maximum: 100000000 }
        monthly_savings: { type: number, exclusiveMinimum: 0, maximum: 100000000, description: Must be less than monthly_take_home. }
        age: { type: integer, minimum: 18, maximum: 80 }
        risk_tolerance: { type: string, enum: [Conservative, Moderate, Aggressive] }
        sold_property_proceeds: { type: number, nullable: true, minimum: 0, maximum: 10000000000, description: Optional historical context only; never deployable capital. Null means not provided. }
        has_lump_sum: { type: boolean, nullable: true, description: Optional. Null means not provided. }
        lump_sum_amount: { type: number, nullable: true, minimum: 0, maximum: 10000000000, description: Positive when has_lump_sum is true, 0 when false, and null when unknown. }
        liquid_savings: { type: number, nullable: true, minimum: 0, maximum: 1000000000, description: Optional; null means unknown and is never treated as zero. }
        emi_burden_pct: { type: number, nullable: true, minimum: 0, maximum: 100, description: Optional; null means unknown and is never treated as zero. }
        financial_dependents: { type: integer, nullable: true, minimum: 0, maximum: 15, description: Optional; null means not provided. }
        emergency_fund_months: { type: number, nullable: true, minimum: 0, maximum: 120, description: Optional; null means unknown and is never treated as zero. }
        investment_goals: { type: array, minItems: 1, maxItems: 4, uniqueItems: true, items: { type: string, enum: [Retirement, Wealth Growth, Tax Saving, Emergency Fund] } }
        investment_horizon_years: { type: integer, minimum: 1, maximum: 30 }
    FinancialProfileCompletion:
      allOf:
        - { $ref: '#/components/schemas/FinancialProfileInput' }
        - type: object
          unevaluatedProperties: false
          properties:
            candidateId: { type: string, format: uuid }
    ProfileCandidate:
      type: object
      required: [candidateId, status, expiresInMs, profileFingerprint, modelVersion, recommendationPolicyVersion, financialProfileSchemaVersion, regulatoryRuleVersion]
      properties:
        candidateId: { type: string, format: uuid }
        status: { type: string, enum: [READY] }
        expiresInMs: { type: integer, minimum: 0 }
        profileFingerprint: { type: string, pattern: '^[a-f0-9]{64}$' }
        modelVersion: { type: string }
        recommendationPolicyVersion: { type: string }
        financialProfileSchemaVersion: { type: string }
        regulatoryRuleVersion: { type: string }
        marketContext: { type: object, additionalProperties: true }
    ProfileCompletion:
      type: object
      required: [profile, recommendation, completion]
      properties:
        profile: { $ref: '#/components/schemas/FinancialProfile' }
        recommendation: { $ref: '#/components/schemas/RecommendationResponse' }
        completion: { type: object, additionalProperties: true }
    FinancialProfileUpdate:
      allOf: [{ $ref: '#/components/schemas/FinancialProfileInput' }, { type: object, required: [version], properties: { version: { type: integer, minimum: 1 } } }]
      unevaluatedProperties: false
    FinancialProfile:
      allOf:
        - { $ref: '#/components/schemas/FinancialProfileInput' }
        - type: object
          required: [profileId, version, monthly_investment_capacity, one_time_investment_capacity, profile_schema_version, risk_capacity_score, risk_capacity_level, final_suitability_level, suitability_reason_codes]
          properties:
            profileId: { type: string, pattern: '^[0-9a-fA-F]{24}$' }
            version: { type: integer, minimum: 1 }
            monthly_investment_capacity: { type: number, minimum: 0 }
            one_time_investment_capacity: { type: number, minimum: 0 }
            profile_schema_version: { type: string }
            risk_capacity_score: { type: number, minimum: 0, maximum: 100 }
            risk_capacity_level: { type: integer, minimum: 1, maximum: 5 }
            final_suitability_level: { type: integer, minimum: 1, maximum: 5 }
            suitability_reason_codes: { type: array, items: { type: string } }
      unevaluatedProperties: false
...
    RecommendationResponse:
      type: object
      required: [recommendationId, audit_id, audit_hash, instruments, model_version, recommendation_policy_version, financial_profile_schema_version, return_basis, final_risk_tier]
      properties: { recommendationId: { type: string }, audit_id: { type: string }, audit_hash: { type: string }, instruments: { type: array, items: { type: object, additionalProperties: true } }, generation_instruments: { type: array, items: { type: object, additionalProperties: true } }, allocation_revision: { type: integer, minimum: 1 }, allocation_revision_id: { type: [string, 'null'] }, portfolio_fingerprint: { type: [string, 'null'], pattern: '^[a-f0-9]{64}$' }, calculation_freshness: { type: object, additionalProperties: true }, state_provenance: { type: object, additionalProperties: true }, confidence_scores: { type: object, additionalProperties: true }, market_adjustment: { $ref: '#/components/schemas/MarketAdjustmentMetadata' }, current_allocation_source: { type: string, enum: [ORIGINAL_RECOMMENDATION, MARKET_CONTEXT_ADJUSTED, USER_REBALANCED] }, model_version: { type: string }, ml_fallback: { type: boolean }, optional_profile_fields_unknown: { type: array, items: { type: string } }, ml_input_fields_unknown: { type: array, items: { type: string } }, recommendation_policy_version: { type: string }, financial_profile_schema_version: { type: string }, return_basis: { type: string, const: PRE_TAX_NOMINAL }, final_risk_tier: { type: string }, excluded_due_to_eligibility: { type: array, items: { type: object, additionalProperties: true } } }
      additionalProperties: true
...
    MarketAdjustmentMetadata:
      type: object
      required: [status, applied, actualTotalTiltPct, reasonCodes, suitabilityValidation, concentrationValidation]
      properties:
        status: { type: string, enum: [APPLIED, NOT_APPLIED, UNAVAILABLE] }
        applied: { type: boolean }
        contextStatus: { type: string }
        context: { type: [string, 'null'] }
        policyVersion: { type: [string, 'null'] }
        observedAt: { type: [string, 'null'], format: date-time }
        evaluatedAt: { type: [string, 'null'], format: date-time }
        actualTotalTiltPct: { type: number, minimum: 0 }
        reasonCodes: { type: array, items: { type: string } }
        suitabilityValidation: { type: string }
        concentrationValidation: { type: string }
        currentAllocationSource: { type: string, enum: [ORIGINAL_RECOMMENDATION, MARKET_CONTEXT_ADJUSTED, USER_REBALANCED] }
        supersededByManualRebalanceAt: { type: [string, 'null'], format: date-time }
      additionalProperties: true
...